requirements.txt
alabaster==1.0.0
anyascii==0.3.2
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
cattrs==23.2.3
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap5==2024.10
cryptography==43.0.1
cssselect2==0.7.0
dealer==2.1.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
Django==4.2.17
django-admin-volt==1.0.14
django-allauth==64.1.0
django-appconf==1.0.6
django-classy-tags==4.1.0
django-cms==4.1.4
django-compressor==4.5.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-downloadview==2.4.0
django-dynamic-preferences==1.16.0
django-entangled==0.5.4
django-environ==0.11.2
django-extensions==3.2.3
django-extra-views==0.14.0
django-filer==3.3.0
django-filter==24.3
djangocms-forms @ git+https://github.com/zite69/djangocms-forms.git@446ad941c79dba86d4ef734310416770a1cfb657
django-formtools==2.5.1
django-fsm==2.8.2
django-haystack==3.3.0
django-image-uploader-widget==1.0.0
django-libsass==0.9
django-localflavor==4.0
django-messages @ git+https://github.com/zite69/django-messages.git@ca7f095ac03e3142d44621c5de31d16106e39e86
django-modelcluster==6.3
django-money==3.5.3
django-oscar==3.2.5
#django-oscar-elasticsearch==2.0.14
django-parler==2.3
django-permissionedforms==0.1
django-phonenumber-field==6.4.0
django-polymorphic==3.1.0
django-sass-processor==1.4.1
django-sekizai==4.1.0
django-select2==8.2.3
django-tables2==2.3.4
django-taggit==5.0.1
django-templates-macros==0.3
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangocms-admin-style==3.3.1
djangocms-alias==2.0.1
djangocms-attributes-field==4.0.0
djangocms-forms @ git+https://github.com/zite69/djangocms-forms.git@446ad941c79dba86d4ef734310416770a1cfb657
djangocms-frontend==1.3.3
djangocms-text-ckeditor==5.1.5
djangocms-versioning==2.0.2
djangorestframework==3.15.2
docutils==0.21.2
draftjs-exporter==5.0.0
easy-thumbnails==2.8.5
elasticsearch==6.8.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
factory-boy==3.2.1
Faker==26.0.0
filetype==1.2.0
html5lib==1.1
idna==3.7
imagesize==1.4.1
jedi==0.19.1
Jinja2==3.1.5
l18n==2021.3
laces==0.1.1
libsass==0.23.0
lsprotocol==2023.0.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
parso==0.8.4
persisting-theory==1.0
pexpect==4.9.0
phonenumbers==8.13.40
phonenumberslite==8.13.42
pillow==11.0.0
pillow_heif==0.20.0
pinax-referrals==4.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
purl==1.6
py-moneyed==3.0
pycountry==24.6.1
pycparser==2.22
pygls==1.3.1
Pygments==2.18.0
pygraphviz==1.13
PyJWT==2.9.0
pyotp==2.9.0
pysolr==3.9.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-stdnum==1.20
pytz==2024.1
rcssmin==1.1.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
rules==3.5
setuptools==75.1.0
six==1.16.0
snowballstemmer==2.2.0
sorl-thumbnail==12.9.0
soupsieve==2.6
spark-parser==1.8.9
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
stack-data==0.6.3
svglib==1.5.1
telepath==0.3.1
tinycss2==1.3.0
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
uwsgidecorators==1.1.0
uwsgidecorators-fallback==0.0.5
#wagtail==6.2
#wagtail-extendedsearch==1.0.7
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.7.0
Willow==1.8.0
xdis==6.1.1
